arm64: mte: Handle panic_on_mte_fault boot parameter
panic_on_mte_fault is a kernel boot option that represents the way on
which the exception handler behaves:
 - off (default) - the exception handler catches only the first fault
                   and prints the error message afterwards it disables
                   the extension.
 - on            - the exception handler generates a panic at the first
                   fault.

Add the code to handle this parameter.

Note: This kernel command line parameter is meant to serve for debugging
purposes. The MTE tag fault, that will be introduced with a future
patch, will be disabled after the first fault since it is meant for
crowd debugging and alternative options might impact performances.
This parameter allows during live debugging sessions to let the kernel
execution stop as a consequence of a tag fault.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>