mm: Add arch_calc_vmalloc_prot_bits to vmalloc
Some architectures might need to calculate the proper page prot based
on the one currently passed to the vmalloc allocator.

This is the case for example when the Memory Tagging Extension is
enabled on arm64, since to get advantage of such a feature PAGE_KERNEL
has to be replaced by PAGE_KERNEL_TAGGED.

Define arch_calc_vmalloc_prot_bits in vmalloc that allows every arch
that redefines it to calculate the correct prototype required to take
advantage of particular architectural features.

The arm64 implementation will be sent with a future patch.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>