arm64: Introduce and use ALTERNATIVE1
We can use a mergable sections for altinstr_replacement, so that
we only get one copy of e.g. "msr pan,#1" instead of 2000.
Use the new macro in the places where it is obvious that there
is one instruction emitted.

Do not bother implementing the same functionality for assembly,
as there are not nearly so many copies as for inline asm.

This reduces the size of defconfig vmlinux by 62k:

 - .altinstr_replacement 00013e14
 + .altinstr_replacement 00004bb0

Signed-off-by: Richard Henderson <richard.henderson@linaro.org>