mm: Introduce the Tag-Based KASAN support for page allocator
The Tag-Based KASAN is a variant of KASAN where the hardware automatically
checks a tag in each pointer against tag metadata assigned to a region of
memory. The implementation of this is architecture-specific.

Introduce the initial implementation of the kasan API for the page
allocator to allow the architectures that expose such a feature to take
advantage of it.

Note: This patch enables only the match-all (0XFF) tag for page access.
The other tags will be enabled in a future patch. The rationale behind
it is that the page allocator is enabled very early in the kernel and at
that point the hw feature required by the Tag-Based KASAN for tagging
purposed might not be enabled yet.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>