mm: Introduce the Tag-Based KASAN for vmalloc
The Tag-Based KASAN is a variant of KASAN where the harware automatically
checks a tag in each pointer against tag metadata assigned to a region of
memory. The implementation of this is architecture-specific.

Introduce the initial implementation of the kasan API for vmalloc to allow
the architecures that expose such a feature to take advantage of it.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>