mm: Introduce the Tag-Based KASAN for SLUB
The Tag-Based KASAN is a variant of KASAN where the hardware automatically
checks a tag in each pointer against tag metadata assigned to a region of
memory. The implementation of this is architecture-specific.

Introduce the initial implementation of the kasan API for SLUB to allow
the architectures that expose such a feature to take advantage of it.

The implementation of the Tag-Based KASAN resembles the one provided by
KASAN_SW. The main differences are:
 * An architecture feature (e.g. arm64 MTE) is required to perform tagging
   operations.
 * Tags are not stored in shadow memory but in a location defined by the
   architecture.
 * The backend implementation is strongly architecture dependent.

Note: This patch introduces some code duplication that it is required
for implementation and bisectability purposes. A future patch will
provide a clean up that removes the duplication. The code that can be
directly reused without changes has been moved into common.c.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>