mm: Enable non match-all tags for page allocator
The page allocator is enabled quite early in the kernel and at that
point the feature required by the Tag-Based KASAN for tagging purposes
might not be available.

The strategy adopted to address this issue is to dispatch non match-all
tags only when the arch feature is detected and enabled.

Introduce an API that allows to the architecture to notify to the Tag-Based
KASAN when the feature has been detected and enabled and use it in the
Tag-Based KASAN implementation for page_alloc to emit correctly the tags.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>