arm64: mte: Implement arch_is_hw_tags_enabled()
The page allocator is enabled quite early in the kernel and at that
point the feature required by the Tag-Based KASAN for tagging purposes
might not be available.

The strategy adopted to address this issue is to dispatch non match-all
tags only when the arch feature is detected and enabled.

Implement the API that allows arm64 to notify to the Tag-Based KASAN when
MTE has been detected and enabled.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>