mm: vmalloc: Sanitize kasan calls for HW features support
KASAN currently uses shadow memory to store the tags that is allocated
at boot and it is populated independently from the memory used to store
data. With the introduction of the Tag-Based KASAN and the consequent
addition of the support for hardware features (e.g. MTE on arm64) that
rely on the underlying memory to store the tags, the approach of writing
tags when the memory is not yet allocated does not work anymore.

Move the kasan calls in places that are sane for all the variants.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>