mm: vmalloc: Fix kasan shadow poisoning size
The size of vm area can be affected by the presence or not of the guard
page. In particular when VM_NO_GUARD is present, the actual accessible
size has to be considered like the real size minus the guard page.

Currently kasan does not keep into account this information during the
poison operation and in particular tries to poison the guard page as
well.

This approach, even if incorrect, does not cause an issue because the
tags for the guard page are written in the shadow memory.
With the introduction of the Tag-Based KASAN, being the guard page
inaccessible by nature, the write tag operation on this page triggers
a fault.

Fix kasan shadow poisoning size invoking get_vm_area_size() instead of
accessing directly the field in the data structure to detect the correct
value.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>