mm: kasan: Update page tag for vmalloc memory
KASAN stores the tag of a page in its flags. When the page is accessed
via the linear mapping this is correctly restored via the page_to_virt()
routine.

Since the memory allocated via vmalloc can be accessed via linear map
the tag stores in struct page needs to be updated accordingly.

Update page tag for vmalloc memory.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>