mm: kasan: Introduce KASAN_VMALLOC_VALID
Currently the KASAN implementation tags as KASAN_VMALLOC_INVALID the
inaccessible allocation and as '0xF' the accessible ones.

Having '0xF' a special meaning for some hardware implementations (e.g.
on ARM64 Memory Tagging Extension aka MTE means match all tags) this choice
might cause issues for the Tag-Based KASAN.

In particular when the chosen backend is MTE the choice allows the
inaccessible locations to be accessible as a side effect.

Introduce KASAN_VMALLOC_VALID and set it to '0x8' to prevent this to
happen.

Note: This patch contains even the required API that allows the
allocations to be tagged correctly.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>