mm: kasan: Enable KASAN_VMALLOC_VALID in vmalloc
The recently introduced KASAN_VMALLOC_VALID allows the Tag-Based KASAN
to detect if a vmalloc'd memory area is accessible (for the purpose of
read and write) or not.

Enable KASAN_VMALLOC_VALID in the vmalloc implementation.

Note: Due to the continuous nature of the per_cpu chunks it is
preferable to allocate a single tag to cover what it is accessible
in the whole area.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>