mm: vmalloc: Reset tag in is_vmalloc_or_module_addr
is_vmalloc_or_module_addr() uses the passed address as an index for
verifying if it belongs to a vmalloc/module allocation or not.

With the introduction of KASAN the address contains a tag that disrupts
such a search.

Reset the address tag before this can be used as an index for the
search.

Note: This patch is required in preparation for the introduction of
KASAN_VMALLOC_VALID.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>