mm: vmalloc: Reset tag in find_vm_area
find_vm_area() uses the passed address as an index for verifying if it
belongs to a specific vm or not.

With the introduction of KASAN the address contains a tag that disrupts
such a search.

Reset the address tag before this can be used as an index for the
search.

Note: This patch is required in preparation for the introduction of
KASAN_VMALLOC_VALID.

signed-off-by: vincenzo frascino <vincenzo.frascino@arm.com>