mm: kasan: Remove vmalloc reserved tags from init set
The initial set of tags that are valid for random generation is passed
by the Tag-Based KASAN to the backend that then makes sure that the
hardware is configured correctly.

A recent patch enabled the kernel to use specific tags for valid/invalid
vmalloc allocations.

Reserve those tags making sure that they are not generated by mistake.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>